Move channel files from pdns/dnsdistdist to pdns
